Add EIP-7702 Check
This PR adds an EIP-7702 support check and warning message if support
cannot be detected. The support works by doing an `eth_call` with a
delegation from a random private key to a state-overriden contract.
This may cause false-negatives (in case the RPC node that is connected
to does not support `eth_call` state overrides for example), so a
"continue anyway" button is included.