Sticky Burner Funds Computation Errors
Computing missing burner funds is a refetching operation, and because
of how `useQuery` works, the error gets cleared in between refetches.
On chains where the burner balance and missing funds fails repeatedly,
this causes the error message to "flash" which isn't great.

Make the error sticky for this refetching query.