Sticky Burner Funds Computation Errors (#23)
Computing missing burner funds is a refetching operation, and because of
how `useQuery` works, the error gets cleared in between refetches. On
chains where the burner balance and missing funds fails repeatedly, this
causes the error message to "flash" which isn't great, and makes it hard
to tell that there is a persistent error in the first place.

Make the error sticky for this refetching query.