Move to Monorepo (#26)
Now that we want to add ERC-7955 tooling (for use with Harbour
specifically), it makes sense to move to a monorepo to prepare for a
hypothetical `hardhat-erc-7955` Hardhat plugin package.

Note that for now, we don't build `@safe-research/erc-7955` and instead
rely on importing TS directly - this is because doing proper CommonJS
and ESM builds would add to this already large PR, and will instead come
in a follow up (when we decide _what_ tooling is actually needed and how
we expect it to be used).

Note most changes were just moving things around and updating paths.
Some other noteworthy changes:
- We re-export the bootstrap artifact with some additional type
annotations.
- Updates to NPM packages (some of them happened when doing monorepo
things, and I thought it made sense to just update them all - as they
are only minor version updates.
- Removed `foundry.toml` (we don't use Foundry anyway)
- Adjusted comment in `Factory.sol` given new paths.

Also, we add explicit scripts to `package.json` instead of relying on
`npm run -ws $SCRIPT` as certain things (like the build) have specific
execution order that is needed for success.