Are We 7955 Yet?
This PR ads a "Are We 7955 Yet?" status page that is accessible from
the main ERC-7955 app. This status page:

1. Fetches the entire ChainList.org database
2. Determines the chain's ERC-7955 status in the background
3. Shows a progress bar with a breakdown of the statuses
4. Searchable chain table

Loading is quite slow (there are over 2000 chains, and MANY of them
time out). Also note that Viem internally handles the retrying (with
exponential backoff) and fallbacking for chains with multiple HTTP RPC
URLs listed (making the code quite a bit simpler).